+++ examples/cnvsparse/demo_cbpdndl_cns_gry_ssd.py
@@ -21,14 +21,13 @@ from sporco import util
 from sporco import plot
 
 
-# Training images (size reduced to speed up demo script)
-exim = util.ExampleImages(scaled=True, zoom=0.25)
-S1 = exim.image('standard', 'lena.grey.png')
-S2 = exim.image('standard', 'barbara.grey.png')
-S3 = util.rgb2gray(exim.image('standard', 'monarch.png',
-                                idxexp=np.s_[:,160:672]))
-S4 = util.rgb2gray(exim.image('standard', 'mandrill.png'))
-S5 = exim.image('standard', 'man.grey.png', idxexp=np.s_[100:612, 100:612])
+# Training images
+exim = util.ExampleImages(scaled=True, zoom=0.25, gray=True)
+S1 = exim.image('barbara.png', idxexp=np.s_[10:522, 100:612])
+S2 = exim.image('kodim23.png', idxexp=np.s_[:, 60:572])
+S3 = exim.image('monarch.png', idxexp=np.s_[:, 160:672])
+S4 = exim.image('sail.png', idxexp=np.s_[:, 210:722])
+S5 = exim.image('tulips.png', idxexp=np.s_[:, 30:542])
 S = np.dstack((S1,S2,S3,S4,S5))
 
 
@@ -45,13 +44,15 @@ D0 = np.random.randn(8, 8, 64)
 
 # Set ConvBPDNDictLearn parameters
 lmbda = 0.2
-opt = cbpdndl.ConvBPDNDictLearn.Options({'Verbose' : True, 'MaxMainIter' : 100,
+opt = cbpdndl.ConvBPDNDictLearn.Options({'Verbose' : True, 'MaxMainIter' : 200,
                                          'CBPDN' : {'rho' : 50.0*lmbda + 0.5},
-                                         'CCMOD' : {'ZeroMean' : True}})
+                                         'CCMOD' : {'rho' : 5,
+                                                    'ZeroMean' : True}},
+                                        method='cns')
 
 
 # Run optimisation
-d = cbpdndl.ConvBPDNDictLearn(D0, sh, lmbda, opt)
+d = cbpdndl.ConvBPDNDictLearn(D0, sh, lmbda, opt, method='cns')
 D1 = d.solve()
 print("ConvBPDNDictLearn solve time: %.2fs" % d.timer.elapsed('solve'))