+++ examples/cnvsparse/demo_cbpdnjnt_clr.py
@@ -21,8 +21,8 @@ import sporco.metric as sm
 
 
 # Load demo image
-img = util.ExampleImages().image('standard', 'barbara.png', scaled=True,
-                                 zoom=0.5)[27:283,55:311]
+img = util.ExampleImages().image('barbara.png', scaled=True, zoom=0.5,
+                                 idxexp=np.s_[10:522, 100:612])
 
 
 # Highpass filter test image