+++ examples/cnvsparse/demo_dictlrn_cbpdn_md.py
@@ -23,18 +23,19 @@ import numpy as np
 
 from sporco.admm import cbpdn
 from sporco.admm import ccmod
+from sporco.admm import ccmodmd
 from sporco.admm import dictlrn
 from sporco import util
 from sporco import plot
 
 
 # Training images (size reduced to speed up demo script)
-exim = util.ExampleImages(scaled=True, zoom=0.25)
-S1 = exim.image('standard', 'lena.grey.png')
-S2 = exim.image('standard', 'barbara.grey.png')
-S3 = util.rgb2gray(exim.image('standard', 'monarch.png'))[:,40:168]
-S4 = util.rgb2gray(exim.image('standard', 'mandrill.png'))
-S5 = exim.image('standard', 'man.grey.png')[25:153, 25:153]
+exim = util.ExampleImages(scaled=True, zoom=0.25, gray=True)
+S1 = exim.image('barbara.png', idxexp=np.s_[10:522, 100:612])
+S2 = exim.image('kodim23.png', idxexp=np.s_[:, 60:572])
+S3 = exim.image('monarch.png', idxexp=np.s_[:, 160:672])
+S4 = exim.image('sail.png', idxexp=np.s_[:, 210:722])
+S5 = exim.image('tulips.png', idxexp=np.s_[:, 30:542])
 S = np.dstack((S1,S2,S3,S4,S5))
 
 
@@ -63,7 +64,7 @@ cri = ccmod.ConvRepIndexing(D0.shape, shp)
 lmbda = 0.2
 optx = cbpdn.ConvBPDNMaskDcpl.Options({'Verbose' : False, 'MaxMainIter' : 1,
                     'rho' : 20.0*lmbda, 'AutoRho' : {'Enabled' : False}})
-optd = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+optd = ccmodmd.ConvCnstrMODMaskDcplOptions({'Verbose' : False,
                     'MaxMainIter' : 1, 'rho' : 2*cri.K,
                     'AutoRho' : {'Enabled' : False}})
 
@@ -84,7 +85,7 @@ xstep = cbpdn.ConvBPDNMaskDcpl(D0n, shp, lmbda, W, optx)
 
 
 # Create D update object
-dstep = ccmod.ConvCnstrMODMaskDcpl(None, shp, W, D0.shape, optd)
+dstep = ccmodmd.ConvCnstrMODMaskDcpl(None, shp, W, D0.shape, optd)
 
 
 # Create DictLearn object