+++ examples/misc/demo_tvl1dcn_dcn_gry.py
@@ -30,8 +30,8 @@ conv = lambda h, x : np.fft.ifft2(np.fft.fft2(h, x.shape)*np.fft.fft2(x)).real
 
 
 # Load reference image
-img = util.rgb2gray(util.ExampleImages().image('standard', 'monarch.png',
-                                               scaled=True))[:,160:672]
+img = util.ExampleImages().image('monarch.png', scaled=True,
+                                 idxexp=np.s_[:,160:672], gray=True)
 
 
 # Construct smoothing filter