+++ examples/misc/demo_tvl2den_clr.py
@@ -21,8 +21,8 @@ from sporco.admm import tvl2
 
 
 # Load reference image
-img = util.ExampleImages().image('standard', 'monarch.png',
-                                 scaled=True)[:,160:672]
+img = util.ExampleImages().image('monarch.png', scaled=True,
+                                 idxexp=np.s_[:,160:672])
 
 
 # Construct test image