+++ sporco/admm/tests/test_ccmod.py
@@ -15,6 +15,59 @@ class TestSet01(object):
 
 
     def test_01(self):
+        N = 32
+        M = 4
+        Nd = 5
+        D0 = ccmod.normalise(ccmod.zeromean(
+            np.random.randn(Nd, Nd, M), (Nd, Nd, M), dimN=2), dimN=2)
+        X = np.zeros((N, N, M))
+        xr = np.random.randn(N, N, M)
+        xp = np.abs(xr) > 3
+        X[xp] = np.random.randn(X[xp].size)
+        S = np.sum(sl.ifftn(sl.fftn(D0, (N, N), (0,1)) *
+                   sl.fftn(X, None, (0,1)), None, (0,1)).real, axis=2)
+        rho = 1e-1
+        opt = ccmod.ConvCnstrMOD_IterSM.Options({'Verbose' : False,
+                    'MaxMainIter' : 500, 'LinSolveCheck' : True,
+                    'ZeroMean' : True, 'RelStopTol' : 1e-5, 'rho' : rho,
+                    'AutoRho' : {'Enabled' : False}})
+        Xr = X.reshape(X.shape[0:2] + (1,1,) + X.shape[2:])
+        Sr = S.reshape(S.shape + (1,))
+        c = ccmod.ConvCnstrMOD_IterSM(Xr, Sr, D0.shape, opt)
+        c.solve()
+        D1 = ccmod.bcrop(c.Y, D0.shape).squeeze()
+        assert(sl.rrs(D0, D1) < 1e-5)
+        assert(np.array(c.getitstat().XSlvRelRes).max() < 1e-5)
+
+
+    def test_02(self):
+        N = 32
+        M = 4
+        Nd = 5
+        D0 = ccmod.normalise(ccmod.zeromean(
+            np.random.randn(Nd, Nd, M), (Nd, Nd, M), dimN=2), dimN=2)
+        X = np.zeros((N, N, M))
+        xr = np.random.randn(N, N, M)
+        xp = np.abs(xr) > 3
+        X[xp] = np.random.randn(X[xp].size)
+        S = np.sum(sl.ifftn(sl.fftn(D0, (N, N), (0,1)) *
+                   sl.fftn(X, None, (0,1)), None, (0,1)).real, axis=2)
+        rho = 1e-1
+        opt = ccmod.ConvCnstrMOD_CG.Options({'Verbose' : False,
+                    'MaxMainIter' : 500, 'LinSolveCheck' : True,
+                    'ZeroMean' : True, 'RelStopTol' : 1e-5, 'rho' : rho,
+                    'AutoRho' : {'Enabled' : False},
+                    'CG' : {'StopTol' : 1e-5}})
+        Xr = X.reshape(X.shape[0:2] + (1,1,) + X.shape[2:])
+        Sr = S.reshape(S.shape + (1,))
+        c = ccmod.ConvCnstrMOD_CG(Xr, Sr, D0.shape, opt)
+        c.solve()
+        D1 = ccmod.bcrop(c.Y, D0.shape).squeeze()
+        assert(sl.rrs(D0, D1) < 1e-4)
+        assert(np.array(c.getitstat().XSlvRelRes).max() < 1e-3)
+
+
+    def test_03(self):
         N = 64
         M = 4
         Nd = 8
@@ -27,27 +80,25 @@ class TestSet01(object):
         S = np.sum(sl.ifftn(sl.fftn(D0, (N, N), (0,1)) *
                    sl.fftn(X, None, (0,1)), None, (0,1)).real, axis=2)
         rho = 1e1
-        opt = ccmod.ConvCnstrMOD.Options({'Verbose' : False,
-                                          'MaxMainIter' : 500,
-                                          'LinSolveCheck' : True,
-                                          'ZeroMean' : True,
-                                          'RelStopTol' : 1e-3, 'rho' : rho,
-                                          'AutoRho' : {'Enabled' : False}})
+        opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
+                    'MaxMainIter' : 500, 'LinSolveCheck' : True,
+                    'ZeroMean' : True, 'RelStopTol' : 1e-3, 'rho' : rho,
+                    'AutoRho' : {'Enabled' : False}})
         Xr = X.reshape(X.shape[0:2] + (1,1,) + X.shape[2:])
         Sr = S.reshape(S.shape + (1,))
-        c = ccmod.ConvCnstrMOD(Xr, Sr, D0.shape, opt)
+        c = ccmod.ConvCnstrMOD_Consensus(Xr, Sr, D0.shape, opt)
         c.solve()
         D1 = ccmod.bcrop(c.Y, D0.shape).squeeze()
         assert(sl.rrs(D0,D1) < 1e-5)
+        assert(np.array(c.getitstat().XSlvRelRes).max() < 1e-5)
 
 
-    def test_02(self):
+    def test_04(self):
         N = 16
         M = 4
         Nd = 8
         X = np.random.randn(N, N, 1, 1, M)
         S = np.random.randn(N, N, 1)
-        opt = ccmod.ConvCnstrMOD.Options({'ZeroMean' : True})
         try:
             c = ccmod.ConvCnstrMOD(X, S, (Nd, Nd, M))
             c.solve()
@@ -56,51 +107,13 @@ class TestSet01(object):
             assert(0)
 
 
-    def test_03(self):
+    def test_05(self):
         N = 16
         M = 8
         X = np.random.randn(N, N, 1, 1, M)
         S = np.random.randn(N, N, 1)
-        opt = ccmod.ConvCnstrMOD.Options({'ZeroMean' : True})
         try:
-            c = ccmod.ConvCnstrMOD(X, S, ((4, 4, 4),(8, 8, 4)), opt)
-            c.solve()
-        except Exception as e:
-            print(e)
-            assert(0)
-
-
-    def test_04(self):
-        N = 16
-        M = 4
-        Nc = 3
-        Nd = 8
-        X = np.random.randn(N, N, Nc, 1, M)
-        S = np.random.randn(N, N, Nc)
-        opt = ccmod.ConvCnstrMOD.Options({'ZeroMean' : True})
-        try:
-            c = ccmod.ConvCnstrMOD(X, S, (Nd, Nd, 1, M), opt, dimK=0)
-            c.solve()
-        except Exception as e:
-            print(e)
-            assert(0)
-
-
-    def test_05(self):
-        N = 16
-        X = np.random.randn(N, N, 3, 1, 10)
-        S = np.random.randn(N, N, 3)
-        dsz = (
-            (
-                (3, 3, 1, 6),
-                (4, 4, 1, 6),
-                (5, 5, 1, 6)
-            ),
-            (6, 6, 3, 4)
-        )
-        opt = ccmod.ConvCnstrMOD.Options({'ZeroMean' : True})
-        try:
-            c = ccmod.ConvCnstrMOD(X, S, dsz, opt, dimK=0)
+            c = ccmod.ConvCnstrMOD(X, S, ((4, 4, 4),(8, 8, 4)))
             c.solve()
         except Exception as e:
             print(e)
@@ -110,11 +123,29 @@ class TestSet01(object):
     def test_06(self):
         N = 16
         M = 4
+        Nc = 3
+        Nd = 8
+        X = np.random.randn(N, N, Nc, 1, M)
+        S = np.random.randn(N, N, Nc)
+        try:
+            opt = ccmod.ConvCnstrMODOptions({'Verbose' : False,
+                            'MaxMainIter' : 20, 'LinSolveCheck' : True})
+            c = ccmod.ConvCnstrMOD(X, S, (Nd, Nd, 1, M), opt=opt, dimK=0)
+            c.solve()
+        except Exception as e:
+            print(e)
+            assert(0)
+        assert(np.array(c.getitstat().XSlvRelRes).max() < 1e-5)
+
+
+    def test_07(self):
+        N = 16
+        M = 4
         Nd = 8
         X = np.random.randn(N, N, 1, 1, M)
         S = np.random.randn(N, N, 1)
         dt = np.float32
-        opt = ccmod.ConvCnstrMOD.Options(
+        opt = ccmod.ConvCnstrMODOptions(
             {'Verbose' : False, 'MaxMainIter' : 20,
              'AutoRho' : {'Enabled' : True},
              'DataType' : dt})
@@ -125,14 +156,14 @@ class TestSet01(object):
         assert(c.U.dtype == dt)
 
 
-    def test_07(self):
+    def test_08(self):
         N = 16
         M = 4
         Nd = 8
         X = np.random.randn(N, N, 1, 1, M)
         S = np.random.randn(N, N, 1)
         dt = np.float64
-        opt = ccmod.ConvCnstrMOD.Options(
+        opt = ccmod.ConvCnstrMODOptions(
             {'Verbose' : False, 'MaxMainIter' : 20,
              'AutoRho' : {'Enabled' : True},
              'DataType' : dt})
@@ -149,11 +180,10 @@ class TestSet01(object):
         Nd = 8
         X = np.random.randn(N, N, 1, 1, M)
         S = np.random.randn(N, N)
-        W = 1.0
         try:
-            opt = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+            opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
                             'MaxMainIter' : 20, 'LinSolveCheck' : True})
-            c = ccmod.ConvCnstrMODMaskDcpl(X, S, W, (Nd, Nd, 1, M),
+            c = ccmod.ConvCnstrMOD_Consensus(X, S, (Nd, Nd, 1, M),
                                            opt=opt, dimK=0)
             c.solve()
         except Exception as e:
@@ -168,11 +198,10 @@ class TestSet01(object):
         Nd = 8
         X = np.random.randn(N, N, 1, K, M)
         S = np.random.randn(N, N, K)
-        W = 1.0
         try:
-            opt = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+            opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
                             'MaxMainIter' : 20, 'LinSolveCheck' : True})
-            c = ccmod.ConvCnstrMODMaskDcpl(X, S, W, (Nd, Nd, 1, M), opt=opt)
+            c = ccmod.ConvCnstrMOD_Consensus(X, S, (Nd, Nd, 1, M), opt=opt)
             c.solve()
         except Exception as e:
             print(e)
@@ -186,11 +215,10 @@ class TestSet01(object):
         Nd = 8
         X = np.random.randn(N, N, Nc, 1, M)
         S = np.random.randn(N, N, Nc)
-        W = 1.0
         try:
-            opt = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+            opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
                             'MaxMainIter' : 20, 'LinSolveCheck' : True})
-            c = ccmod.ConvCnstrMODMaskDcpl(X, S, W, (Nd, Nd, 1, M),
+            c = ccmod.ConvCnstrMOD_Consensus(X, S, (Nd, Nd, 1, M),
                                            opt=opt, dimK=0)
             c.solve()
         except Exception as e:
@@ -206,11 +234,10 @@ class TestSet01(object):
         Nd = 8
         X = np.random.randn(N, N, Nc, K, M)
         S = np.random.randn(N, N, Nc, K)
-        W = 1.0
         try:
-            opt = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+            opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
                             'MaxMainIter' : 20, 'LinSolveCheck' : True})
-            c = ccmod.ConvCnstrMODMaskDcpl(X, S, W, (Nd, Nd, 1, M),
+            c = ccmod.ConvCnstrMOD_Consensus(X, S, (Nd, Nd, 1, M),
                                            opt=opt)
             c.solve()
         except Exception as e:
@@ -226,11 +253,10 @@ class TestSet01(object):
         Nd = 8
         X = np.random.randn(N, N, Nc, K, M)
         S = np.random.randn(N, N, Nc, K)
-        W = 1.0
         try:
-            opt = ccmod.ConvCnstrMODMaskDcpl.Options({'Verbose' : False,
+            opt = ccmod.ConvCnstrMOD_Consensus.Options({'Verbose' : False,
                             'MaxMainIter' : 20, 'LinSolveCheck' : True})
-            c = ccmod.ConvCnstrMODMaskDcpl(X, S, W, (Nd, Nd, Nc, M),
+            c = ccmod.ConvCnstrMOD_Consensus(X, S, (Nd, Nd, Nc, M),
                                            opt=opt)
             c.solve()
         except Exception as e:
@@ -239,7 +265,14 @@ class TestSet01(object):
 
 
     def test_14(self):
-        opt = ccmod.ConvCnstrMOD.Options({'AuxVarObj' : False})
+        opt = ccmod.ConvCnstrMODBase.Options({'AuxVarObj' : False})
         assert(opt['fEvalX'] is True and opt['gEvalY'] is False)
         opt['AuxVarObj'] = True
         assert(opt['fEvalX'] is False and opt['gEvalY'] is True)
+
+
+    def test_15(self):
+        opt = ccmod.ConvCnstrMODBase.Options({'AuxVarObj' : True})
+        assert(opt['fEvalX'] is False and opt['gEvalY'] is True)
+        opt['AuxVarObj'] = False
+        assert(opt['fEvalX'] is True and opt['gEvalY'] is False)